requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.1.10 3.2.16
django-cors-headers==3.14.0
djangorestframework==3.11.2
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1